tbank-emo/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.0
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.5
openai==1.77.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0